White Paper PDF file with development details
Only the methodology section was included in the White Paper due to privacy reasons. The sections of teammates will be included with their appropriate permission.

The methodology section contains information regarding the development of the Damn Vulnerable Android Application - an overview, used languages and environment, base application and each vulnerability. Only the Data Leakage vulnerability is included in the repo - the rest are missing due to time constraints or lack of data from teammates which had to handle the development for the specific vulnerability.