Data Leak Documentation
Thorough explanation of the vulnerability, a full walkthrough of the challenge within the application and appropriate countermeasures.

The vulnerability was found when I examined the application's generated APK file then thoroughly analysed and researched through a variety of articles and papers. All sources are referenced within the file White Paper PDF (ZenTech_DVAA.pdf).